pyproject.toml
[project]
name = "gold-split-examples"
version = "0.1.0"
description = "Examples using the GoldSplitter from Goldener during training Machine Learning models"
readme = "README.md"
requires-python = ">=3.10"
license-files= ["LICENSE"]
authors = [
    {name = "goldener-data", email = "yann.chene.tni@gmail.com"},
]
keywords = ["machine-learning", "data-splitting", "goldener", "pytorch", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch",
    "lightning",
    "mlflow",
    "goldener",
    "scikit-learn",
    "torchmetrics",
    "hydra-core",
]

[project.optional-dependencies]
vision = [
    "torchvision",
    "pillow"
]

dev = [
    "ruff",
    "mypy",
    "pre-commit",
]


[tool.uv]
dev-dependencies = [
    "ruff",
    "mypy",
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/goldener-data/gold-splitter-examples"
Repository = "https://github.com/goldener-data/gold-splitter-examples"
Issues = "https://github.com/goldener-data/gold-splitter-examples/issues"
